5_huggingface/requirements.txt
transformers==4.57.6
hf_transfer==0.1.9
accelerate==1.12.0
diffusers==0.36.0
datasets==3.6.0
faiss-cpu==1.11.0
matplotlib==3.10.3